+++ .gitignore
@@ -6,3 +6,6 @@
 /old movies.sqlite3
 /movies 2.sqlite3
 /identifier.sqlite
+/pytest.ini
+/.idea/sqldialects.xml
+/test/tmdb/.coverage
